Fallback method refactor (#197)
* refactor theme support options

* Move the theme support files to single directory

* Fix header and footer templates should include required actions

* Add force width css for the header
