+++ inc/class-header-footer-elementor.php
@@ -50,8 +50,6 @@ class Header_Footer_Elementor {
 	 */
 	function __construct() {
 
-		$hfe_compatibility_option = get_option( 'hfe_all_theme_support_option', '1' );
-
 		$this->template = get_template();
 
 		if ( defined( 'ELEMENTOR_VERSION' ) && is_callable( 'Elementor\Plugin::instance' ) ) {
@@ -77,17 +75,8 @@ class Header_Footer_Elementor {
 
 				require HFE_DIR . 'themes/oceanwp/class-hfe-oceanwp-compat.php';
 			} else {
+				require_once HFE_DIR . 'themes/default/class-hfe-fallback-theme-support.php';
 
-				add_action( 'admin_menu', array( $this, 'hfe_settings_page' ) );
-				add_action( 'admin_init', array( $this, 'hfe_save_setting_data' ) );
-				if ( '1' === $hfe_compatibility_option ) {
-					add_action( 'get_header', array( $this, 'override_header' ) );
-					add_action( 'get_footer', array( $this, 'override_footer' ) );
-					require HFE_DIR . 'themes/default/class-hfe-default-compat.php';
-				} elseif ( '2' === $hfe_compatibility_option ) {
-					add_action( 'get_header', array( $this, 'option_override_header' ) );
-					require HFE_DIR . 'themes/global-theme-fallback/class-global-theme-compatibility.php';
-				}
 				add_action( 'init', array( $this, 'setup_unsupported_theme_notice' ) );
 			}
 
@@ -109,104 +98,6 @@ class Header_Footer_Elementor {
 			add_action( 'network_admin_notices', array( $this, 'elementor_not_available' ) );
 		}
 
-	}
-
-	/**
-	 * Function for overriding the header in the elmentor way.
-	 *
-	 * @since 1.0.16
-	 *
-	 * @return void
-	 */
-	public function override_header() {
-		require HFE_DIR . 'themes/default/hfe-header.php';
-		$templates   = array();
-		$templates[] = 'header.php';
-		// Avoid running wp_head hooks again.
-		remove_all_actions( 'wp_head' );
-		ob_start();
-		locate_template( $templates, true );
-		ob_get_clean();
-	}
-
-	/**
-	 * Function for overriding the footer in the elmentor way.
-	 *
-	 * @since 1.0.16
-	 *
-	 * @return void
-	 */
-	public function override_footer() {
-		require HFE_DIR . 'themes/default/hfe-footer.php';
-		$templates   = array();
-		$templates[] = 'footer.php';
-		// Avoid running wp_head hooks again.
-		remove_all_actions( 'wp_head' );
-		ob_start();
-		locate_template( $templates, true );
-		ob_get_clean();
-	}
-
-	/**
-	 * Function overriding the header in the wp_body_open way.
-	 *
-	 * @since 1.0.16
-	 *
-	 * @return void
-	 */
-	public function option_override_header() {
-		$templates   = array();
-		$templates[] = 'header.php';
-		locate_template( $templates, true );
-		if ( ! did_action( 'wp_body_open' ) ) {
-			echo '<div class="force-stretched-header">';
-			do_action( 'hfe_fallback_header' );
-			echo '</div>';
-		}
-	}
-
-	/**
-	 * Show a settings page incase of unsupported theme.
-	 *
-	 * @since 1.0.16
-	 *
-	 * @return void
-	 */
-	public function hfe_settings_page() {
-		add_submenu_page(
-			'options-general.php',
-			'Header Footer Elementor',
-			'Header Footer Elementor',
-			'manage_options',
-			'hfe',
-			array( $this, 'hfe_settings_page_html' )
-		);
-	}
-	/**
-	 * Save the data from the settings page.
-	 *
-	 * @since 1.0.16
-	 *
-	 * @return void
-	 */
-	public function hfe_save_setting_data() {
-		// bail if our option is not set.
-		if ( empty( $_POST['hfe_radio_button'] ) ) {
-			return;
-		}
-
-		$hfe_radio = sanitize_text_field( wp_unslash( $_POST['hfe_radio_button'] ) );
-		update_option( 'hfe_all_theme_support_option', $hfe_radio );
-	}
-	/**
-	 * Settings page.
-	 *
-	 * Settings page markup in the file which is included.
-	 *
-	 * @since 1.0.0
-	 */
-	public function hfe_settings_page_html() {
-		require_once HFE_DIR . 'inc/hfe-settings-page.php';
 	}
 
 	/**
