+++ themes/default/class-hfe-default-compat.php
@@ -11,13 +11,6 @@
 class HFE_Default_Compat {
 
 	/**
-	 * Instance of HFE_Default_Compat.
-	 *
-	 * @var HFE_Default_Compat
-	 */
-	private static $instance;
-
-	/**
 	 * Instance of Elementor Frontend class.
 	 *
 	 * @var \Elementor\Frontend()
@@ -27,15 +20,12 @@ class HFE_Default_Compat {
 	/**
 	 *  Initiator
 	 */
-	public static function instance() {
-		if ( ! isset( self::$instance ) ) {
-			self::$instance = new HFE_Default_Compat();
-			add_action( 'wp', array( self::$instance, 'hooks' ) );
-		}
+	public function __construct() {
 		if ( defined( 'ELEMENTOR_VERSION' ) && is_callable( 'Elementor\Plugin::instance' ) ) {
 			self::$elementor_instance = Elementor\Plugin::instance();
 		}
-		return self::$instance;
+
+		add_action( 'wp', [ $this, 'hooks' ] );
 	}
 
 	/**
@@ -43,28 +33,62 @@ class HFE_Default_Compat {
 	 */
 	public function hooks() {
 		if ( hfe_header_enabled() ) {
+			// Replace header.php template.
+			add_action( 'get_header', array( $this, 'override_header' ) );
+
+			// Display HFE's header in the replaced header.
 			add_action( 'hfe_header', 'hfe_render_header' );
 		}
 
 		if ( hfe_footer_enabled() ) {
+			// Replace footer.php template.
+			add_action( 'get_footer', array( $this, 'override_footer' ) );
+
+			// Display HFE's footer in the replaced header.
 			add_action( 'hfe_footer', 'hfe_render_footer' );
 		}
 
 		if ( hfe_is_before_footer_enabled() ) {
-			// Action `elementor/page_templates/canvas/after_content` is introduced in Elementor Version 1.9.0.
-			if ( version_compare( ELEMENTOR_VERSION, '1.9.0', '>=' ) ) {
-				// check if current page template is Elemenntor Canvas.
-				if ( 'elementor_canvas' == get_page_template_slug() ) {
-					$override_cannvas_template = get_post_meta( hfe_get_before_footer_id(), 'display-on-canvas-template', true );
-					if ( '1' == $override_cannvas_template ) {
-						add_action( 'elementor/page_templates/canvas/after_content', array( 'Header_Footer_Elementor', 'get_before_footer_content' ), 9 );
-					}
-				}
-			}
 			add_action( 'hfe_footer_before', array( 'Header_Footer_Elementor', 'get_before_footer_content' ) );
 		}
 	}
 
+	/**
+	 * Function for overriding the header in the elmentor way.
+	 *
+	 * @since 1.0.16
+	 *
+	 * @return void
+	 */
+	public function override_header() {
+		require HFE_DIR . 'themes/default/hfe-header.php';
+		$templates   = array();
+		$templates[] = 'header.php';
+		// Avoid running wp_head hooks again.
+		remove_all_actions( 'wp_head' );
+		ob_start();
+		locate_template( $templates, true );
+		ob_get_clean();
+	}
+
+	/**
+	 * Function for overriding the footer in the elmentor way.
+	 *
+	 * @since 1.0.16
+	 *
+	 * @return void
+	 */
+	public function override_footer() {
+		require HFE_DIR . 'themes/default/hfe-footer.php';
+		$templates   = array();
+		$templates[] = 'footer.php';
+		// Avoid running wp_head hooks again.
+		remove_all_actions( 'wp_head' );
+		ob_start();
+		locate_template( $templates, true );
+		ob_get_clean();
+	}
+
 }
 
-HFE_Default_Compat::instance();
+new HFE_Default_Compat();
