+++ themes/default/hfe-footer.php
@@ -8,7 +8,6 @@
 
 ?>
 
-
 <?php do_action( 'hfe_footer' ); ?>
 <?php wp_footer(); ?>
 </body>
