+++ themes/default/hfe-header.php
@@ -17,6 +17,7 @@
 </head>
 
 <body <?php body_class(); ?>>
+<?php do_action( 'wp_body_open' ); ?>
 <div id="page" class="hfeed site">
 
 <?php do_action( 'hfe_header' ); ?>
